requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.4
pyttsx3==2.90
matplotlib==3.8.4
python-dotenv==1.0.1